requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.0
firecrawl-py==2.7.1
frozenlist==1.6.2
google-search-results==2.4.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipython==8.37.0
jedi==0.19.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.64
langchain-openai==0.3.21
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
marshmallow==3.26.1
matplotlib-inline==0.1.7
multidict==6.4.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.2.6
openai==1.84.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.51
propcache==0.3.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.41
stack-data==0.6.3
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
